core: add "nm-sd-utils.h" to access system internal helper
We have a fork of a lot of useful systemd helper code.
However, until now we shyed away from using it aside from
the bits that we really need.

That means, we need to re-implement this.

Add "nm-sd-utils.h" header to very carefully make internal
systemd API accessible to the rest of core.

Use with caution.