Query and ping does not require root
Change permission of the daemon API socket to 0666 allowing anyone to
send message to daemon.

In daemon, only ping and query is allowed for non-root user. Other
commands will trigger error for non-root user. In the future, we will
make sure secrets not pass to query by non-root also, since we do not
have such secret in schema yet, no action required for this patch.

The implementation is just use `getsockopt(SO_PEERCRED)` to get the peer
UID of socket connection.

Signed-off-by: Gris Ge <fge@redhat.com>